Reverted the file change to use partial.v8. We decided it was too confusing having a separate file for just a few of the services.  Due to all the other changes that were committed, decided to recommit this one file instead of reverting the previous commit. Removed the Obsolete warning from the NotNullOrWhitespace function.
