+++ Rock/Utility/ExtensionMethods/StringExtensions.cs
@@ -114,8 +114,6 @@ namespace Rock
         /// <returns>
         ///   <c>true</c> if [is not null or whitespace] [the specified string]; otherwise, <c>false</c>.
         /// </returns>
-        [RockObsolete( "1.8" )]
-        [Obsolete( "Use IsNotNullOrWhiteSpace instead. Fixes non-standard casing.", false )]
         public static bool IsNotNullOrWhitespace( this string str )
         {
             return !string.IsNullOrWhiteSpace( str );
